Move tests around for 2.13 stable
