+++ testing/test_package/pubspec.yaml
@@ -8,4 +8,4 @@ dependencies:
   test_package_imported:
     path: "../test_package_imported"
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.13.0 <3.0.0'
